feat: implement PDF generation endpoint
